Half Hour 7
Made levelCheck() dynamic by checking if the lastLevel is lesser than the current level and checking if the most left wall (to the left of the player) exists, then reset the doors. The lastLevel check is used so that when going from level 2 to level 1, it won't reset the doors.